change group number
According to josh, group #131 is no longer in use for determining if a contact is a contributor or active contributor.  Instead group #487 has been used, which includes 6 child groups.  I'm changing the group id# so the link should appear (hopefully) on contributor pages such as https://civicrm.org/user/####/edit